int/int64-cast

+++ opensource/kernel-cpp/scheduler/spooler.cxx
@@ -2032,7 +2032,7 @@ string Spooler::configuration_for_single_job_script()
 void Spooler::start()
 {
     static_log_categories.save_to( &_original_log_categories );
-    _max_micro_step_time = Duration(_variables->get_int64("scheduler.message.SCHEDULER-721.timeout", _max_micro_step_time.seconds()));
+    _max_micro_step_time = Duration((double)_variables->get_int64("scheduler.message.SCHEDULER-721.timeout", _max_micro_step_time.seconds()));
 
     _state_cmd = sc_none;
     set_state( s_starting );
